+++ app/scripts/__init__.py
@@ -1,1 +1,0 @@
-"""package containing scripts that are essential for the running of BCC"""