+++ start_bcc.sh
@@ -77,15 +77,6 @@ do
     done
 done
 
-# Load configuration files, passing the command-line args to load_config
-# for now load_config expects only one option: "--device DEVICE_FILE_PATH"
-# other possible arguments in $@ will be passed back as "unrecognized_args",
-# so they can be used if needed, in later parts of this script
-unrecognized_args=$(python -m app.scripts.load_config "$@")
-
-if test -n "$unrecognized_args"; then
-   echo "Arguments unrecognized by load_config.py: $unrecognized_args, in case needed below"
-fi
 
 # Worker processes
 rq worker "${DEFAULT_PREFIX}_job_registration" &